refactor: move tile-related events to `World.Events` and update event handlers
- Relocated `ThingAddedToTileEvent`, `ThingRemovedFromTileEvent`, and `ThingUpdatedOnTileEvent` to `World.Events` for better domain alignment.
- Updated corresponding event handlers and namespaces to reflect the changes.
- Added missing contracts for `Creatures` across multiple files to consolidate dependencies.
