+++ src/ApplicationServer/NeoServer.Server.Events/Subscribers/EventSubscriber.cs
@@ -15,7 +15,6 @@ using NeoServer.Server.Events.Items;
 using NeoServer.Server.Events.Player;
 using NeoServer.Server.Events.Player.Trade;
 using NeoServer.Server.Events.Server;
-using NeoServer.Server.Events.Tiles;
 using NeoServer.Server.Services;
 
 namespace NeoServer.Server.Events.Subscribers;
