+++ src/ApplicationServer/NeoServer.Server.Routines/Creatures/Monster/MonsterDefenseRoutine.cs
@@ -1,4 +1,5 @@
-﻿using NeoServer.Server.Common.Contracts;
+﻿using NeoServer.Domain.Common.Contracts.Creatures;
+using NeoServer.Server.Common.Contracts;
 using NeoServer.Server.Tasks;
 
 namespace NeoServer.Server.Routines.Creatures.Monster;
