+++ src/ApplicationServer/NeoServer.Server.Routines/Creatures/Monster/MonsterStateRoutine.cs
@@ -1,4 +1,5 @@
-﻿using NeoServer.Domain.Creatures.Monster.Services;
+﻿using NeoServer.Domain.Common.Contracts.Creatures;
+using NeoServer.Domain.Creatures.Monster.Services;
 
 namespace NeoServer.Server.Routines.Creatures.Monster;
