+++ src/Core/NeoServer.Domain/Common/Contracts/Creatures/IMonster.cs
@@ -1,11 +1,11 @@
-﻿using NeoServer.Domain.Common.Contracts.Creatures;
-using NeoServer.Domain.Common.Contracts.Services;
+﻿using NeoServer.Domain.Common.Contracts.Services;
 using NeoServer.Domain.Common.Contracts.World;
 using NeoServer.Domain.Common.Item;
-using NeoServer.Domain.Common.Location.Structs;
 using NeoServer.Domain.Creatures;
 using NeoServer.Domain.Creatures.Monster;
 using NeoServer.Domain.Creatures.Monster.Combat;
+
+namespace NeoServer.Domain.Common.Contracts.Creatures;
 
 public interface IMonster : IWalkableMonster, ICombatActor
 {
@@ -73,7 +73,7 @@ public interface IMonster : IWalkableMonster, ICombatActor
     void UpdateState();
 
     void Escape();
-    void Born(Location location);
+    void Born(Location.Structs.Location location);
     void CreateSummon(ISummonService summonService);
     void PostAttack(MonsterCombatType type);
     bool IsImmune(Immunity immunity);
