+++ src/Core/NeoServer.Domain/Creatures/Events/Monster/MonsterStateChangedEvent.cs
@@ -1,4 +1,5 @@
 using NeoServer.Domain.Common;
+using NeoServer.Domain.Common.Contracts.Creatures;
 using NeoServer.Domain.Creatures.Monster;
 
 namespace NeoServer.Domain.Creatures.Events.Monster;
