+++ src/Core/NeoServer.Domain/Creatures/Events/Monster/MonsterWasBornEvent.cs
@@ -1,4 +1,5 @@
 using NeoServer.Domain.Common;
+using NeoServer.Domain.Common.Contracts.Creatures;
 using NeoServer.Domain.Common.Location.Structs;
 
 namespace NeoServer.Domain.Creatures.Events.Monster;
