+++ src/Core/NeoServer.Domain/World/Events/CreatureAddedOnMapEvent.cs
@@ -2,7 +2,7 @@ using NeoServer.Domain.Common;
 using NeoServer.Domain.Common.Contracts.Creatures;
 using NeoServer.Domain.Common.Contracts.World;
 
-namespace NeoServer.Domain.Creatures.Events;
+namespace NeoServer.Domain.World.Events;
 
 /// <summary>
 /// Event raised when a walkable creature is added to the map.
