+++ src/Core/NeoServer.Domain/World/Events/ThingAddedToTileEvent.cs
@@ -2,7 +2,7 @@ using NeoServer.Domain.Common;
 using NeoServer.Domain.Common.Contracts.Items;
 using NeoServer.Domain.Common.Contracts.World;
 
-namespace NeoServer.Domain.Items.Events;
+namespace NeoServer.Domain.World.Events;
 
 /// <summary>
 /// Event raised when a thing (item) is added to a tile.
