+++ src/Extensions/NeoServer.Scripts.LuaJIT/Functions/MonsterFunctions.cs
@@ -1,4 +1,5 @@
 ﻿using LuaNET;
+using NeoServer.Domain.Common.Contracts.Creatures;
 using NeoServer.Scripts.LuaJIT.Enums;
 using NeoServer.Scripts.LuaJIT.Functions.Interfaces;
 using NeoServer.Server.Common.Contracts;
