+++ src/NetworkingServer/NeoServer.Networking/EventHandlers/World/CreatureAddedOnMapEventHandler.cs
@@ -3,24 +3,17 @@ using NeoServer.Domain.Common.Contracts.Creatures;
 using NeoServer.Domain.Common.Contracts.World;
 using NeoServer.Domain.Common.Creatures;
 using NeoServer.Domain.Common.Helpers;
-using NeoServer.Domain.Creatures.Events;
+using NeoServer.Domain.World.Events;
 using NeoServer.Networking.Packets.Outgoing.Creature;
 using NeoServer.Networking.Packets.Outgoing.Effect;
 using NeoServer.Networking.Packets.Outgoing.Item;
 using NeoServer.Server.Common.Contracts;
 using NeoServer.Server.Common.Contracts.Network;
 
-namespace NeoServer.Server.Events.Creature;
+namespace NeoServer.Networking.EventHandlers.World;
 
-public class CreatureAddedOnMapEventHandler : INetworkingEventHandler<CreatureAddedOnMapEvent>
+public class CreatureAddedOnMapEventHandler(IGameServer game) : INetworkingEventHandler<CreatureAddedOnMapEvent>
 {
-    private readonly IGameServer game;
-
-    public CreatureAddedOnMapEventHandler(IGameServer game)
-    {
-        this.game = game;
-    }
-
     public void Handle(CreatureAddedOnMapEvent @event)
     {
         Execute(@event.Creature, @event.Cylinder);
