+++ src/NetworkingServer/NeoServer.Networking/EventHandlers/World/Tiles/ThingAddedToTileEventHandler.cs
@@ -3,12 +3,12 @@ using NeoServer.Domain.Common.Contracts.Creatures;
 using NeoServer.Domain.Common.Contracts.Items;
 using NeoServer.Domain.Common.Contracts.World;
 using NeoServer.Domain.Common.Helpers;
-using NeoServer.Domain.Items.Events;
+using NeoServer.Domain.World.Events;
 using NeoServer.Networking.Packets.Outgoing.Item;
 using NeoServer.Server.Common.Contracts;
 using NeoServer.Server.Common.Contracts.Scripts;
 
-namespace NeoServer.Server.Events.Tiles;
+namespace NeoServer.Networking.EventHandlers.World.Tiles;
 
 public class ThingAddedToTileEventHandler(IGameServer game, IScriptManager scriptManager)
     : INetworkingEventHandler<ThingAddedToTileEvent>
