+++ tests/NeoServer.Domain.Tests/Creature/Experience/ProportionalExperienceModifierTest.cs
@@ -1,5 +1,6 @@
 ﻿using Moq;
 using NeoServer.Domain.Common.Combat;
+using NeoServer.Domain.Common.Contracts.Creatures;
 using NeoServer.Domain.Creatures.Experience;
 using NeoServer.Domain.Tests.Helpers.Player;
