+++ tests/NeoServer.Domain.Tests/World/TileLoadTests.cs
@@ -5,6 +5,7 @@ using NeoServer.Domain.Common.Location.Structs;
 using NeoServer.Domain.Items.Events;
 using NeoServer.Domain.Tests.Helpers;
 using NeoServer.Domain.Tests.Helpers.Map;
+using NeoServer.Domain.World.Events;
 using NeoServer.Domain.World.Map;
 using NeoServer.Domain.World.Models.Tiles;
